more optimizations

+++ tt_npe/cpp/include/npeCommon.hpp
@@ -13,7 +13,7 @@ namespace tt_npe {
 using BytesPerCycle = float;
 using Cycle = uint64_t;
 using Timestep = uint64_t;
-using DeviceID = int16_t;
+using DeviceID = uint32_t;
 #define MESH_DEVICE -1
 
 enum class nocType { NOC0 = 0, NOC1 = 1 };
